requirements.txt
PyQt5>=5.15.0
python-vlc>=3.0.0
paramiko>=2.7.0